Split the access and spine tiers out of eval.rs
